Promote from create-test-user to main (#83)
* Feat: Pipeline Playwright Tests

Add Rest Api Client ID to env vars

Refs: BLAIS5-2974

* Update rest-api-node-client

* 200 ms is far too short in CI

* WIP: Create and delete test user

Functionality written to create and delete test user however test is failing when trying to install an instrument - despite already having created a user and connected to the Rest Api. Potential timeout issue?

Refs: BLAIS5-2974

* WIP: Create and delete test user

Functionality written to create and delete test user however test is failing - definitely a timeout issue.

Refs: BLAIS5-2974

* feat: Create and delete test user

Commit includes creation and deletion of test user as well as new log in to MIR functionality. This commit works but code is fugly! Requires some serious refactoring.

Refs: BLAIS5-2974

* Refactor: Create and delete test user

Debug log statements added

Refs: BLAIS5-2974

* Refactor: Create and delete test user

Abstracted helpers from primary test module

Refs: BLAIS5-2974

* Fix: Create and delete test user

Update dependencies

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Keeping all globals centralised.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Further abstractions and DRYing-ish.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Further abstractions and DRYing-ish.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Shuffled some signatures and renamed some functions.

Refs: BLAIS5-2974

* Debug: Create and delete test user.

Remove teardown to see how far Concourse gets.

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Tests are passing locally.

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Increase timeout and add debug logs.

Refs: BLAIS5-2974

* Feat: Create and delete test user.

Added functionality to log out of MIR as part of the teardown.

Refs: BLAIS5-2974

* Debug: Create and delete test user.

I've run out of theories. Seriously. Removing the teardown because I don't believe it should be failing :rage:

Refs: BLAIS5-2974

* Debug: Create and delete test user.

Re-implementing teardown.

Refs: BLAIS5-2974

* Fi: Create and delete test user.

For the love of God I hope setting the locale fixes it! :pray:

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Increase timeout because CATI...

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Increase timeout again because CATI...

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Increase timeout AGAIN because CATI...

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Increase page.click timeout...

Refs: BLAIS5-2974

* Fix: Create and delete test user.

Doubled page.click timeout...

Refs: BLAIS5-2974

* Use IAP for blaise api client on cleanup

* Refactor: Create and delete test user.

Removed function to logout of MIR, removed debug logs, and removed some unnecessary parameters.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Messed about with some globals.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Decreased timeout where possible and removed unused parameter.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Made it look beautiful with the team.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Whitespacing.

Refs: BLAIS5-2974

* Refactor: Create and delete test user.

Make some small specifications within the README.

Refs: BLAIS5-2974

* update dependencies

* Upgrade blaise-design-system
