+++ package.json
@@ -31,7 +31,8 @@
     "@types/react-datepicker": "^3.1.8",
     "@types/react-dom": "^17.0.2",
     "axios": "^0.21.0",
-    "blaise-design-system-react-components": "git+https://github.com/ONSdigital/blaise-design-system-react-components",
+    "blaise-api-node-client": "https://github.com/ONSdigital/blaise-api-node-client",
+    "blaise-design-system-react-components": "https://github.com/ONSdigital/blaise-design-system-react-components",
     "blaise-iap-node-provider": "https://github.com/ONSdigital/blaise-iap-node-provider",
     "dayjs": "^1.10.6",
     "dotenv": "^8.2.0",
@@ -72,7 +73,6 @@
     "@typescript-eslint/eslint-plugin": "^4.17.0",
     "@typescript-eslint/parser": "^4.16.1",
     "axios-mock-adapter": "^1.19.0",
-    "blaise-api-node-client": "https://github.com/ONSdigital/blaise-api-node-client",
     "concurrently": "^7.0.0",
     "cross-env": "^7.0.2",
     "eslint": "^7.21.0",
